Return index type changed

+++ src/Main.java
@@ -3,7 +3,7 @@ public class Main {
         BmiService service = new BmiService();
         double heightInMeters = 1.78;
         int weightInKilo = 75;
-        int bmi = (int) service.calculate(heightInMeters, weightInKilo);
+        int bmi = service.calculate(heightInMeters, weightInKilo);
         System.out.println("Рост: " + heightInMeters);
         System.out.println("Вес: " + weightInKilo);
         System.out.println("Индекс массы тела: " + bmi);
